pyproject.toml
[project]
name = "SnapTeX"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keyboard>=0.13.5",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
]

